Switched to selenium; added a source for adserver hostnames
Selenium will open an instance of the Firefox browser when given a URL,
and will wait until the page is completely loaded to grab the page’s
HTML source. Thus, any JavaScript on the page that changes the HTML to
show an ad will fire, and then we can grab the HTML afterward.

requests.get() doesn’t seem to work the same way.

Also, I found a site made by a guy who is really fed up with ads, and
he maintains a 2000+ long list of hostnames that serve ads:
https://pgl.yoyo.org/as/. I’ve added some code that grabs the plaintext
version of the list of hostnames, and then generates a list to compare
the links that show up in the HTML source.

From here, I think it’s time to play around with regex to clean up the
formatting of the lists and allow for easy matching.